partitioned workers and fixed most bugs
