+++ app.rb
@@ -11,9 +11,9 @@ end
 
 get "/service/:service" do
    if(params[:service] === "google")
-       system("ruby text-track-service.rb #{params[:service]} /D/innovation/text-track-service test <put google auth file here> bbb-accessibility")
+       system("ruby text-track-service.rb #{params[:service]} /D/innovation/text-track-service test <put auth file path> bbb-accessibility")
    elsif(params[:service] === "ibm")
-       system("ruby text-track-service.rb #{params[:service]} /D/innovation/text-track-service test <put apikey here>")
+       system("ruby text-track-service.rb #{params[:service]} /D/innovation/text-track-service test <put api key here>")
    else
        puts "No such service found"
    end
